pyproject.toml
[project]
name = "ai-agent-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.12.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
